requirements.txt
tqdm
numpy
pandas
matplotlib
scikit-learn
pytorch
pycocotools